General structure setup - moving from PHP/IIS to NODE/LINUX
