--- AutoBlender/pyBlender/Scripts/SceneSetup.txt
@@ -1,92 +1,0 @@
-import bpy, math, os, sys
-from math import sin, cos, pi
-from mathutils import Vector, Quaternion
-import logging, time
-
-time.sleep(2)
-logging.basicConfig(
-		filename = "Out.txt",
-		level = logging.DEBUG, 
-		format = '[pyBlender]: %(message)s'
-	)
-
-def sout(input):
-	logging.info(input)
-
-# let the user know Blender is working
-#
-sout("")
-sout("")
-sout("     ATTN: Blender is now working!      ")
-sout("----------------------------------------")
-sout("")
-sout("")
-
-# open and load the beat times information
-#
-beatData = []
-beatFile = open('.\\pyBlender\\Output\\beat_frames.lfa','r')
-for line in beatFile.readlines():
-    beatData.extend(line.split())
-beatFile.close()
-
-beatFrames = []
-beats = 0
-for i in range(len(beatData)):
-    tmp = beatData[i]
-    beatFrames.append(int(float(tmp)))
-
-sout("Beats Finished Importing")
-sout("Beats: " + str(len(beatData)))
-sout("")
-time.sleep(1)
-
-# open and load the mel volume information
-#
-volumeData = []
-volumeFile = open('.\\pyBlender\\Output\\mel_volume.lfa','r')
-for line in volumeFile.readlines():
-    volumeData.extend(line.split())
-volumeFile.close()
-
-# open and load mel information
-#
-MelData = []
-melFile = open('.\\pyBlender\\Output\\n_bin_mel.lfa','r')
-for line in melFile.readlines():
-    MelData.append(line.split())
-melFile.close()
-
-maxFrames = len(MelData)
-melBins = len(MelData[0])
-
-sout("Mel Finished Importing")
-sout("Mel Frames: " + str(maxFrames))
-sout("Mel Bins: " + str(melBins))
-sout("")
-time.sleep(1)
-
-# setup some basic parameters before creating/animating objects
-#
-scn = bpy.context.scene
-scn.frame_end = maxFrames
-
-origin = Vector().zero()
-meshLoc = Vector().zero()
-
-frameSkip = int(sys.argv[8])
-
-sout("Frames traveled per keyframe: " + str(frameSkip))
-sout("")
-sout("")
-sout("")
-sout("***** ANIMATION STARTED *****")
-sout("")
-
-#
-#	ATTENTION: THIS ENDS THE SCENE SETUP AREA
-#	
-#	THE REST OF THE TOOL CHAIN IS SIMPLY A RENDERING SCRIPT, AKA "CORE"
-#	AND CAN NOW BE REPLACED, MODIFIED, OR REIMPLIMENTED AT WILL
-#
-#
