--- AutoBlender/terminator.php
@@ -1,4 +1,0 @@
-<?php
-$nuke = fopen("terminate", 'c+');
-fclose($nuke);
-?>
